[refactor] Use camelCase for crypto
