+++ docs/@aptos-labs/ts-sdk-0.0.0/classes/Signature.html
@@ -30,7 +30,7 @@ message and can be used with the PublicKey to verify the signature.</p>
 <li><a href="MultiEd25519Signature.html" class="tsd-signature-type" data-tsd-kind="Class">MultiEd25519Signature</a></li>
 <li><a href="Secp256k1Signature.html" class="tsd-signature-type" data-tsd-kind="Class">Secp256k1Signature</a></li></ul></li></ul></li></ul></section><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L65">src/core/crypto/asymmetric_crypto.ts:65</a></li></ul></aside>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L65">src/core/crypto/asymmetricCrypto.ts:65</a></li></ul></aside>
 <section class="tsd-panel-group tsd-index-group">
 <section class="tsd-panel tsd-index-panel">
 <details class="tsd-index-content tsd-index-accordion" open><summary class="tsd-accordion-summary tsd-index-summary">
@@ -99,7 +99,7 @@ This function is the Typescript SDK equivalent of <code>bcs::to_bytes</code> in 
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">void</span></h4><aside class="tsd-sources">
 <p>Overrides <a href="Serializable.html">Serializable</a>.<a href="Serializable.html#serialize">serialize</a></p>
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L76">src/core/crypto/asymmetric_crypto.ts:76</a></li></ul></aside></li></ul></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L76">src/core/crypto/asymmetricCrypto.ts:76</a></li></ul></aside></li></ul></section>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="toString" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagAbstract">Abstract</code> <span>to<wbr/>String</span><a href="#toString" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <ul class="tsd-signatures tsd-kind-method tsd-parent-kind-class">
@@ -109,7 +109,7 @@ This function is the Typescript SDK equivalent of <code>bcs::to_bytes</code> in 
 </div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">string</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L74">src/core/crypto/asymmetric_crypto.ts:74</a></li></ul></aside></li></ul></section>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L74">src/core/crypto/asymmetricCrypto.ts:74</a></li></ul></aside></li></ul></section>
 <section class="tsd-panel tsd-member tsd-kind-method tsd-parent-kind-class"><a id="toUint8Array" class="tsd-anchor"></a>
 <h3 class="tsd-anchor-link"><code class="tsd-tag ts-flagAbstract">Abstract</code> <span>to<wbr/>Uint8<wbr/>Array</span><a href="#toUint8Array" aria-label="Permalink" class="tsd-anchor-icon"><svg class="icon icon-tabler icon-tabler-link" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><use href="#icon-anchor-a"></use><use href="#icon-anchor-b"></use><use href="#icon-anchor-c"></use></svg></a></h3>
 <ul class="tsd-signatures tsd-kind-method tsd-parent-kind-class">
@@ -119,7 +119,7 @@ This function is the Typescript SDK equivalent of <code>bcs::to_bytes</code> in 
 </div>
 <h4 class="tsd-returns-title">Returns <span class="tsd-signature-type">Uint8Array</span></h4><aside class="tsd-sources">
 <ul>
-<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L69">src/core/crypto/asymmetric_crypto.ts:69</a></li></ul></aside></li></ul></section></section></div>
+<li>Defined in <a href="https://github.com/aptos-labs/aptos-ts-sdk/blob/6081872/src/core/crypto/asymmetric_crypto.ts#L69">src/core/crypto/asymmetricCrypto.ts:69</a></li></ul></aside></li></ul></section></section></div>
 <div class="col-4 col-menu menu-sticky-wrap menu-highlight">
 <div class="tsd-navigation settings">
 <details class="tsd-index-accordion"><summary class="tsd-accordion-summary">
