+++ src/core/authentication_key.ts
@@ -3,9 +3,9 @@
 
 import { sha3_256 as sha3Hash } from "@noble/hashes/sha3";
 import { AccountAddress } from "./account_address";
-import { PublicKey } from "./crypto/asymmetric_crypto";
+import { PublicKey } from "./crypto/asymmetricCrypto";
 import { Ed25519PublicKey } from "./crypto/ed25519";
-import { MultiEd25519PublicKey } from "./crypto/multi_ed25519";
+import { MultiEd25519PublicKey } from "./crypto/multiEd25519";
 import { Secp256k1PublicKey } from "./crypto/secp256k1";
 import { Hex } from "./hex";
 import { AuthenticationKeyScheme, HexInput, SigningScheme } from "../types";
