+++ src/core/crypto/index.ts
@@ -1,7 +1,7 @@
 // Copyright © Aptos Foundation
 // SPDX-License-Identifier: Apache-2.0
 
-export * from "./asymmetric_crypto";
+export * from "./asymmetricCrypto";
 export * from "./ed25519";
-export * from "./multi_ed25519";
+export * from "./multiEd25519";
 export * from "./secp256k1";
