+++ src/transactions/types.ts
@@ -6,7 +6,7 @@ import { MoveObject, MoveOption, MoveString, MoveVector } from "../bcs/serializa
 import { Bool, U128, U16, U256, U32, U64, U8 } from "../bcs/serializable/movePrimitives";
 import { FixedBytes } from "../bcs/serializable/fixedBytes";
 import { AccountAddress } from "../core";
-import { PublicKey } from "../core/crypto/asymmetric_crypto";
+import { PublicKey } from "../core/crypto/asymmetricCrypto";
 import {
   MultiAgentRawTransaction,
   FeePayerRawTransaction,
